+++ Source/Server/Source/Modules/Objects/ServerQuest.vb
@@ -6,7 +6,7 @@ Imports ASFW.IO.FileIO
 Friend Module ServerQuest
 #Region "Constants"
     'Constants
-    Friend Const MAX_QUESTS As Byte = 250
+    Friend Const MAX_QUESTS As Integer = 250
     'Friend Const MAX_TASKS As Byte = 10
     'Friend Const MAX_REQUIREMENTS As Byte = 10
     Friend Const MAX_ACTIVEQUESTS = 10
